Move origin private file system support to a separate module
