+++ src/index.ts
@@ -4,13 +4,12 @@ export { WASI, WASIProcExit };
 export { Fd, Inode } from "./fd.js";
 export {
   File,
-  SyncOPFSFile,
   Directory,
   OpenFile,
   OpenDirectory,
-  OpenSyncOPFSFile,
   PreopenDirectory,
   ConsoleStdout,
 } from "./fs_core.js";
+export { SyncOPFSFile, OpenSyncOPFSFile } from "./fs_opfs.js";
 export { strace } from "./strace.js";
 export * as wasi from "./wasi_defs.js";
